API Integration
Replaced all the fake data with a real API call. Instead of looping through the days, we are looping through the 'forecast' variable (which is an array of 8 items for each upcoming day). Then we rounded the maximum and minimum temperature. Then we only displayed 5 days, instead of 8, by adding an If statement.